backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
google-genai==1.7.0
pydantic==2.10.6
websockets==15.0.1
